+++ jestConfig.js
@@ -4,6 +4,9 @@ module.exports = {
     'config/webpack/*.js',
     'scripts/aggregate-translations/**/*.js',
   ],
+  setupFiles: [
+    './jestsetup.js',
+  ],
   coverageDirectory: 'tests/jest/reports/coverage',
   coverageReporters: [
     'html',
@@ -15,5 +18,8 @@ module.exports = {
     '**/jest/**/(*.)(spec|test).js?(x)',
   ],
   roots: [process.cwd()],
+  snapshotSerializers: [
+    './node_modules/enzyme-to-json/serializer',
+  ],
   testURL: 'http://localhost',
 };